agent/requirements.txt
python-socketio==5.9.0
python-engineio==4.7.1
psutil==5.9.5
requests==2.31.0
python-dotenv==1.0.0
PyInstaller==6.1.0